Remove nexpose subpackage and remove unused files
There should only be one domain for this project, so the nexpose
package is not needed. It was removed and the corresponsing changes
(nexpose.Asset to domain.Asset) were made as a result. Common.go and
Vulnerability.go are also no longer needed since the work that would
need that is being split off into another service that hydrates assets.