Add settings library to read in and use environment variables
We made a library (https://github.com/asecurityteam/settings) to help
manage configuration our services might need. For the nexpose-vuln-notifier,
we need Nexpose configuration to know how to connect to the Nexpose instance
and hit the API to get Nexpose Assets, and we need Producer information to know
the endpoint we should produce the Nexpose Assets to. This change adds that
configuration.