Fix imports and naming, update README
We renamed the repo from nexpose-vuln-notifier to
nexpose-asset-producer because we changed the scope of the project
to only produce the assets instead of fully hydrating the assets with
vulnerability details and solutions and let the latter be accomplished
by another service. Renaming the repo means that docs, imports, files,
and variable names needed to be updated with the new name. The README
was also updated to include more information about the service.