Don't produce empty assets.
This change adds a check to see if the asset that's recieved in the
assetChannel is empty. Although we don't expect to have empty assets,
this can occur if Nexpose returns an empty asset or an asset that is
missing the ID, IP, or the last time it was scanned.