Remove SiteAssetsResponse.Links field since it's not used
The Links that Nexpose sends back in its response is not needed, so it
is being removed in this change. All we need from Nexpose here are
are the Assets (Resources) so that we can produce them for the next
consumer and the Page in order to get the total number pages we will
need to paginate through and the total number of assets that will be
returned.